requirements_hosting.txt
# Requirements mínimos para hosting
# Optimizado para evitar conflictos de dependencias

# Framework web principal
Flask>=2.3.0,<3.0.0
Werkzeug>=2.3.0,<3.0.0

# Base de datos
pymysql>=1.0.0

# Configuración
python-dotenv>=1.0.0

# Para el proxy HTTP MySQL
flask-cors>=4.0.0
requests>=2.30.0

# Manejo de archivos Excel (opcional)
pandas>=2.0.0
openpyxl>=3.1.0

# Utilidades básicas
psutil>=5.9.0
pytz>=2023.3